+++ docs/sections/changelog.rst
@@ -8,6 +8,11 @@ All notable changes to this project will be documented in this file.
 Unreleased
 ==========
 
+* Added: token instrospection endpoint support (RFC7662).
+* Added: request in password grant authenticate call.
+* Changed: dropping support for Django versions before 1.8.
+* Changed: pass token and request to OIDC_IDTOKEN_PROCESSING_HOOK.
+
 0.6.0
 =====